[feature/CRUD] get MAP API
