+++ src/components/mainPages/MakeDatingCourse.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function MakeDatingCourse() {
+  return (
+    <div>
+      <div>데이트코스</div>
+      <div>장소1</div>
+      <div>장소2</div>
+      <div>장소3</div>
+      <button>저장</button>
+    </div>
+  );
+}
+
+export default MakeDatingCourse;
