+++ src/pages/DetailPages.jsx
@@ -1,7 +1,21 @@
-import React from "react";
+import React from 'react';
 
 function DetailPages() {
-  return <div>DetailPages</div>;
+  return (
+    <div>
+      <div>맵</div>
+      <div>
+        실제 저장한 장소 1 <button>길찾기</button>
+      </div>
+      <div>
+        실제 저장한 장소 2<button>길찾기</button>
+      </div>
+      <div>
+        실제 저장한 장소 3<button>길찾기</button>
+      </div>
+      <button>공유하기</button>
+    </div>
+  );
 }
 
 export default DetailPages;
