+++ src/pages/MainPages.jsx
@@ -1,13 +1,16 @@
+import MakeDatingCourse from 'components/mainPages/MakeDatingCourse';
+import CompletedDatingCourse from 'components/mainPages/CompletedDatingCourse';
 import MapSearch from 'components/mainPages/MapSearch';
 import SearchResult from 'components/mainPages/SearchResult';
 import React from 'react';
 
 function MainPages() {
   return (
-    <div>
+    <>
       <MapSearch />
       <SearchResult />
-    </div>
+      <CompletedDatingCourse />
+    </>
   );
 }
